add Main Menu and Pause Menu
This Commit contains changes that include the first versions of
Game Main Menu Scene, as well as Pause Menu in the gameplay.
So far most of changes here are just placeholders.

1. Main Menu Scene with that can Play the game (by running  Scene "1-1")
2. Settings are implemented, although those are just placeholders,
    I would need to still put in the Volume, and Brightness
3. Quit button with Quit Menu to it
 Additionally
1. Pause menu with even more functions fabricated as an asset,
    to be simply plug and play in the scene
2. All the above functions are integrated into that menu

What needs to be done?
1. I need someone @Daniele Carro @Kohei Tateyama @Makoto Mezawa of you to
 show me how to disable all input buttons when i press the Pause Menu button
(top-right on the video). I could do it myself,
but its gonna be much much faster with a help of someone who did it for dialogues.
2. Connect settings functionalities to actual things.
3. Add save and load menu and things connected to it.