feat: add controls for musicSource in different scenes
This is a squash commit that integrates all the previous changes done in this feature.
We have added Music Togge on and off for music, Music Slider for control and a Manager script for music.

This is a combination of 4 commits.
This is the 1st commit message:

feat: Freeze time when pause menu is on & return to main menu

Added two functions in main manager for handling Time.scale() to make sure
we freeze the game while the pause menu is on.

This is the commit message #2:

feat: Music Toggle & Music Slider control the audio source + add Music Manager

1. Music Slider controls the volume of music from 0 to  100%
2. Music Toggle Stops the music from being played, and replays it when turned on again
3. All those actions are done by Music Manager that is persisten across all the scenes,
this component actually controls the audio source. Also if no audio source exists,
it creates new audio source to surpress errors from showing up in certain scenes, such as Main Menu

to make this work changes to the scenes had to be applied. Most of the functions are implemented in .cs files,
however, we have added prefabs for music manager to make sure it is the same across all the scenes.

Next step in development will be to add the control of screen brightness,
and creating a template for all the scenes

This is the commit message #3:

fix: Make MusicManager and Toggle work together

The issue was corresponding with errors that were showing up when Music
Manager instance was recreated and connected to Toggle and Slider. Also
There was an issue of having lack of the AudioSource.

For that we have made musicManager being recreated in the scene if called
by another gameObject.

At this point this solution doesn't show flaws,
but might not be the best overall.

 This is the commit message #4:

fix: Make MusicSlider work without audio source